penjualan add

+++ application/views/penjualan/add.php
@@ -43,37 +43,24 @@
           <table id="input-penjualan" class="table">
             <thead class="bg-secondary text-white">
               <tr>
-                <th style="width: 200px;">Barang</th>
-                <th style="width: 170px;">Harga</th>
-                <th style="width: 80px;">Qty</th>
-                <th style="width: 80px;">Tonase</th>
-                <th style="width: 80px;">Diskon %</th>
-                <th style="width: 150px;">Total</th>
+                <th>Barang</th>
+                <th>Harga</th>
+                <th style="width: 100px;">Qty</th>
+                <th style="width: 100px;">Diskon %</th>
+                <th style="width: 200px;">Total</th>
                 <th class="text-right">
-                  <button onclick="addItem()" type="button" class="btn btn-info btn-sm">
+                  <button onclick="pilihBarang()" type="button" class="btn btn-info btn-sm">
                     <i class="fa fa-plus"></i>
                   </button>
                 </th>
               </tr>
             </thead>
             <tbody>
-              <tr id="baris-0">
-                <td>
-                  <select style="width: 200px;" name="idbarang[]" class="form-control idbarang" type="text">
-                    <option>Pilih</option>
-                  </select>
-                </td>
-                <td><input id="harga_jual-0" name="harga_jual[]" style="width: 170px;" type="text" placeholder="Harga" class="form-control harga_jual"></td>
-                <td><input value="0" id="qty-0" name="qty[]" type="number" style="width: 70px;" class="form-control qty"></td>
-                <td><input value="0" id="tonase-0" name="tonase[]" type="number" style="width: 70px;" class="form-control tonase"></td>
-                <td><input value="0" id="diskon-0" name="diskon[]" type="text" style="width: 70px;" class="form-control diskon"></td>
-                <td id="total-0">0</td>
-                <td class="text-right"></td>
-              </tr>
+
             </tbody>
             <tfoot>
               <tr>
-                <td colspan="4"></td>
+                <td colspan="3"></td>
                 <td span="">Total Harga</td>
                 <td id="total-semua">0</td>
                 <td></td>
@@ -88,7 +75,46 @@
   </div>
 </div>
 
+
+<div class="modal fade" id="modal_form" data-backdrop="static" tabindex="-1" role="dialog">
+  <div class="modal-dialog modal-lg modal-dialog-scrollable" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <h5 class="modal-title" id="modal_title">Pilih Barang</h5>
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div class="modal-body form">
+        <form action="#" id="form" class="form-horizontal">
+          <div class="form-body">
+            <div class="form-row">
+              <div class="col-md-12 col-xs-12">
+                <div class="form-group">
+                  <input name="q" placeholder="Cari" class="form-control" type="search">
+                </div>
+              </div>
+            </div>
+          </div>
+          <div id="listbarang" class="row col-md-12">
+          </div>
+        </form>
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div>
+
 <script>
+  $(document).ready(function() {
+    $.ajax({
+      url: '<?= base_url('ajax/get_barang') ?>',
+      method: 'GET',
+      success: function(data) {
+        $('#listbarang').html(data);
+      }
+    });
+  })
+
   function selectBarang() {
     $('.idbarang').select2({
       theme: 'bootstrap4',
@@ -149,27 +175,40 @@
     });
   }
 
-  function addItem() {
-    let baris = parseInt($('#input-penjualan > tbody tr:last').attr('id').split('-')[1]);
-    console.log(baris)
-    $('tbody').append(`<tr id="baris-${baris+1}">
-                <td> <select style="width: 200px;" name="idbarang[]" class="form-control idbarang" type="text">
-                    <option>Pilih</option>
-                  </select></td>
-                <td><input id="harga_jual-${baris+1}" name="harga_jual[]" style="width: 170px;" type="text" placeholder="Harga" class="form-control harga_jual"></td>
-                <td><input id="qty-${baris+1}" value="0" name="qty[]" type="number" style="width: 70px;" class="form-control qty"></td>
-                <td><input id="tonase-${baris+1}" value="0" name="tonase[]" type="number" style="width: 70px;" class="form-control tonase"></td>
-                <td><input id="diskon-${baris+1}" value="0" name="diskon[]" type="text" style="width: 70px;" class="form-control diskon"></td>
-                <td id="total-${baris+1}">0</td>
+  function pilihBarang() {
+    $('#modal_form').modal('show')
+  }
+
+  function addItem(item) {
+
+    var harga = new Intl.NumberFormat("id-ID", {
+      style: "currency",
+      currency: "IDR"
+    }).format(parseFloat(item.harga_jual));
+    let as = $(`#baris-${item.idbarang}`).attr('id')
+    if (as) {
+      return alert('sudah ada')
+    }
+    $('tbody').append(`<tr id="baris-${item.idbarang}">
+                <td>
+                  <div class="">
+                    <p class="m-0">${item.nmbarang} | ${item.merek}</p>
+                    <small class="text-secondary">${item.nmsatuan}</small><br>
+                    <small class="text-secondary">${item.nmkategori}</small><br>
+                  </div>
+                </td>
+                <td id="harga_jual-${item.idbarang}">
+                ${harga}
+                </td>
+                <td><input value="0" id="qty-${item.idbarang}" name="qty[]" type="number" class="form-control qty"></td>
+                <td><input value="0" id="diskon-${item.idbarang}" name="diskon[]" type="text" class="form-control diskon"></td>
+                <td id="total-${item.idbarang}">0</td>
                 <td class="text-right"><button onclick="deleteItem(this)" type="button" class="btn btn-sm btn-danger"><i class="fa fa-minus"></i></button></td>
               </tr>`)
     selectBarang()
     checkInputMask()
   }
 
-  $('body').on('input', '.harga_jual', function() {
-    hitungRow(this)
-  });
   $('body').on('input', '.qty', function() {
     hitungRow(this)
   });
@@ -179,7 +218,7 @@
 
   function hitungRow(element) {
     var baris = $(element).attr('id').split('-')[1];
-    var price = parseFloat($('#harga_jual-' + baris).val().replace('.', '')) || 0;
+    var price = parseFloat(Number($('#harga_jual-' + baris).text().replace(/,.*|[^0-9]/g, '')))
     var qty = parseInt($('#qty-' + baris).val()) || 0;
     var diskon = parseInt($('#diskon-' + baris).val()) || 0;
     var total = price * qty;
@@ -194,10 +233,10 @@
 
   function hitung() {
     let total = 0;
-    $('#input-penjualan > tbody  > tr').each(function(i) {
-      total += parseFloat(Number($('#total-' + i).text().replace(/,.*|[^0-9]/g, '')))
+    $('#input-penjualan > tbody  > tr').each(function() {
+      var baris = $(this).attr('id').split('-')[1]
+      total += parseFloat(Number($('#total-' + baris).text().replace(/,.*|[^0-9]/g, '')))
     });
-    console.log()
     var totalrp = new Intl.NumberFormat("id-ID", {
       style: "currency",
       currency: "IDR"
